mean and acm are float values

+++ include/rpiraw.h
@@ -54,11 +54,11 @@
                                     const unsigned hist_width,
                                     const unsigned hist_height);
 
-    uint32_t rpiraw_calc_mean_by_hist(uint32_t *hist, const unsigned len,
-                                      const unsigned npixs);
+    float rpiraw_calc_mean_by_hist(uint32_t *hist, const unsigned len,
+                                   const unsigned npixs);
 
-    uint32_t rpiraw_calc_acm_by_hist(uint32_t *hist, const unsigned len,
-                                     const unsigned npixs, const uint32_t mean);
+    float rpiraw_calc_acm_by_hist(uint32_t *hist, const unsigned len,
+                                  const unsigned npixs, const uint32_t mean);
 
     int rpiraw_histogram_equalize_rgb888(uint8_t *dst, const unsigned ld_dst,
                                          uint8_t *src, const unsigned ld_src,
