+++ src/statistic.c
@@ -17,8 +17,8 @@
  *   Thirty-Sixth Southeastern Symposium on System Theory,
  *   2004. Proceedings of the, 2004, pp. 472-475.
  */
-uint32_t rpiraw_calc_mean_by_hist(uint32_t *hist, const unsigned len,
-                                  const unsigned npixs)
+float rpiraw_calc_mean_by_hist(uint32_t *hist, const unsigned len,
+                               const unsigned npixs)
 {
     unsigned i;
     uint_fast32_t sum = 0;
@@ -27,7 +27,7 @@ uint32_t rpiraw_calc_mean_by_hist(uint32_t *hist, const unsigned len,
     for (i = 0; i < len; i ++)
         sum += i * hist[i];
 
-    return sum / npixs;
+    return (float) sum / npixs;
 }
 
 /*
@@ -35,8 +35,8 @@ uint32_t rpiraw_calc_mean_by_hist(uint32_t *hist, const unsigned len,
  *   Thirty-Sixth Southeastern Symposium on System Theory,
  *   2004. Proceedings of the, 2004, pp. 472-475.
  */
-uint32_t rpiraw_calc_acm_by_hist(uint32_t *hist, const unsigned len,
-                                 const unsigned npixs, const uint32_t mean)
+float rpiraw_calc_acm_by_hist(uint32_t *hist, const unsigned len,
+                              const unsigned npixs, const uint32_t mean)
 {
     unsigned i;
     uint_fast32_t sum = 0;
@@ -45,5 +45,5 @@ uint32_t rpiraw_calc_acm_by_hist(uint32_t *hist, const unsigned len,
     for (i = 0; i < len; i ++)
         sum += abs((int32_t) i - mean) * hist[i];
 
-    return sum / npixs;
+    return (float) sum / npixs;
 }
